+++ server/libs/sqlite/util.js
@@ -1,0 +1,68 @@
+/** ****************************************************************************
+ *
+ * A ...
+ *
+ * util.js is just a literal object that contains a set of functions.
+ * It can't be instantiated.
+ *
+ * Private Functions:
+ *  . ...                         ...,
+ *
+ *
+ * Public Static Methods:
+ *  . ...                         ...,
+ *
+ *
+ *
+ * @namespace    -
+ * @dependencies none
+ * @exports      -
+ * @author       -
+ * @since        0.0.0
+ * @version      -
+ * ************************************************************************** */
+/* eslint-disable no-underscore-dangle */
+
+
+// -- Vendor Modules
+
+
+// -- Local Modules
+
+
+// -- Local Constants
+
+
+// -- Local Variables
+
+
+// -- Private Functions --------------------------------------------------------
+
+
+// 2020-10-13 18:27:26
+function _getDateTime(date) {
+  const year = date.getFullYear();
+  const month = date.getMonth() < 9 ? `0${date.getMonth() + 1}` : date.getMonth() + 1;
+  const day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate();
+
+  const hours = date.getHours() < 10 ? `0${date.getHours()}` : date.getHours();
+  const min = date.getMinutes() < 10 ? `0${date.getMinutes()}` : date.getMinutes();
+  const sec = date.getSeconds() < 10 ? `0${date.getSeconds()}` : date.getSeconds();
+
+  return `${year}-${month}-${day} ${hours}:${min}:${sec}`;
+}
+
+
+// -- Public Static Methods ----------------------------------------------------
+
+const Util = {
+  getDateTime(date) {
+    return _getDateTime(date);
+  },
+};
+
+
+// -- Export
+module.exports = Util;
+
+/* - */